Refactor commands into their own package; #261
